Use bg appop ALLOWED, not DEFAULT
DEFAULT is not actually the right thing; it's a subtlety used for
other things that have compound state.  ALLOWED is the proper
outright "let this thing run free" state for this appop.

Also, set the appop only for the target package, not for the
entire uid.

Bug 37226920
Bug 37242946
Test: make RunSettingsRoboTests

Change-Id: I8c0a995b9519f980685e8573da946dcc93ac4e7c
(cherry picked from commit 65cb0d648885101abbe317ffb3ae8c2771384485)