VPN settings: better a11y voice for legacy dialog
Previously, the user input fields in the legacy VPN config dialog would
simply be read out as "edit box", "drop-down list", etc., without
reference to their corresponding labels. This is not optimal for
low-vision users.

In this CL, we add a "labelFor" attribute to the prompt TextView of each
field. The resulting input fields would instead be read out as "edit box
for username", "drop-down list for type", etc., which are much less
ambiguous than before.

Bug: 37202471
Test: make SettingsGoogle and navigate the dialog with TalkBack on
Change-Id: Id23d0563984e3de4a4a3d03a587ca130b50508fc