Fix setting crash when removing screenlock with fp
If a phone is operating slower than normal, when removing screen lock,
it's possible for the user to go back to settings page while removal
is still in progress. The removal callback will crash ChooseLockGeneric

Fixes: 37329856
Test: in fpc_tee_hal.c, make do_remove slow (5s). set up lockscreen, fingerprint,
then set lockscreen to none or swipe. do device rotations, go back to lock screen,
and then try to enroll fingerprint

Change-Id: I6da131a0001b22d06e359401f2ca1c6496c071ca