Only show headers in Master Clear when there are multiple profiles
The Master Clear dialog lists the accounts whose data will be lost by
a factory reset. The list contains the headings "Personal" and "Work"
to distinguish accounts in the main user and the work profile. When
there is no work profile, the header (only "Personal" would be shown in
this case) is superfluous and misleading (as the user may be using the
device for work).

Bug: 30132270
Test: Manual :(

Change-Id: Id493652fc331d4500c4d772a9abcc716995565b3