Turn off QUIET_MODE before changing unified lock credential.
For FBE-enabled devices, the work profile disk encryption key is evicted
when the the profile is turned off (i.e. QUIET_MODE is on). If the user
attempts to change the profile's lock screen credentials in this state,
the profile will get into an unrecoverable state as vold does not have
the current encryption key.

Currently, separate work challenge does not suffer from this issue as
QUIET_MODE must be turned off before the change credentials flow could
be initiated in Settings. In this CL, we fix the issue for unified work
challenge as well by adopting the same approach.

This is temporary and should be removed when vold no longer relies on
the current key for resetting the password.

Bug: 35930129
Test: manually resetting unified PIN when work mode is off
      on both sailfish (FBE) and angler (no FBE)
Test: make SettingsRoboTests

Change-Id: Ibb4b1c40a9cad8a473e08cf0c294699406108cd9