Fix an override filter bug.
The override filter for storage to ensure that we filter by UUID and
user id was designed as an override filter. This worked as long as there
never was any other filters which could be used out of Storage Settings.
It turns out the instant apps filter applies here.

As a result, I've moved these filters into being "composite filters"
which are composed with the primary filter instead of overriding the
primary filter.

Fixes: 35245772
Test: Settings Robotests... but as for the composing part it is
b/34768986

Change-Id: I6f5acb909568ab02e81187c09fdda9533d635781