Revert "Revert "Revert "Replace dynamic summary text for SettingPref"""
This reverts commit c1e9f8787f02df0d38a9d4e93f7996318fc46355.

Reason for revert: There is an issue when dropdown menu text has a '%' sign in it.
Despite testing it on my own devices, the bug did not show up immediately.

Instead of continuing down this path for OC, we'll just revert
this patch, and then as an ugly, band-aid solution, check the
summary text for the "%s" substring, and hide it when found.
In O-DR onward, we'll look for a more permanent solution.

Change-Id: Ia7a04bfb7b09c436dc0f13564f2134a1f0c436e2
Bug: 36101902
Fix: 62211676
(cherry picked from commit ed761c4e743bcad72c23e2d37f479474850a42c8)