Delay showing the loading progress spinner in Managed App.
In warm load, the app list is populated much faster, and we should not
show the spinner if we can show the app list immediately.

Adding 100ms delay in showing the loading container. If the app list is
ready, show the list container directly without first showing the
loading container.

Change-Id: I8b3ea88d2382ece9226d5c7734fd1aa406160cf2
Fix: 36375109
Test: make RunSettingsRoboTests