Adjust UX for toggles when active scorer is not set.
This change fixes two issues: when wakeup is not available, update the
number of items to show above the fold in wifi preferences. When scorer
is not set, grey out wakeup toggle and set summary to reflect the
status.

Bug: 62410973
Bug: 62343859
Test: make RunSettingsRoboTests -j40
Merged-In: I5f4131c6a86eab6c0fb03ea2b5101ba107a9189f
Change-Id: I5f4131c6a86eab6c0fb03ea2b5101ba107a9189f