Add location as a Inline list payload
The inline payload is added to security settings because
security settings indexes the setting with the key
"location" which corresponds with the correct
PreferenceController, LocationPreferenceController.

However, we suppress the Location search result, so to
avoid confusion later, we don't include the ResultPayload
in the search indexing.

Bug: 62022517
Test: make RunSettingsRoboTests
Change-Id: Ica9b775fe7b43259e651f95b079b8cccb29d0ea8
Merged-In: Ic5515d38064498ae908f4cefedfeb3c2a2af449c