Fix bug in DevicePickerFragment
BluetoothProgressCategory will add a no device preference if it is
not scanning and no device is added to the list. Since
DeviceListFragment doesn't report its scanning status correctly, it will
show "No nearby device" even though it is scanning.

This cl add control code for BluetoothProgressCategory in
DeviceListFragment, which avoid the unnecessary no device preference.

Bug: 62492822
Test: RunSettingsRoboTests
Change-Id: Ic49f21cb58a80dcd37c5737800181263ae876ebc