Fix crash when dismissing suggestions.
When the suggestions are in default mode and the number of suggestions
are more than the default count, we set the sublist of the suggestions
list as the suggestion adapter data. In this case, operation on the
suggestions list should be done in the sublist instead of the original
list to maintain the list consistency check.

Hence, when dismissing the suggestion, change to remove the suggestion
from the sugestion adapter instead of from the dashboard adapter, so
that any following interaction with the suggestions will not trigger the
concurrent modification exception.

Change-Id: I4caba74cfcc7bd0d53ad8c7dffcb127b3ebd845d
Merged-In: I970b6af8a1c72bc0c3ee89fef890ae6a669c71d2
Fix: 64279080
Test: make RunSettingsRoboTests