Update availability for mobile network
This CL updates the availability to behave properly when a
user other than the primary user is accessing it. This also
makes the api return no intent if they don't have permission to
access this because attempting to go to that screen will cause a
crash. Also updates tests related to this change.

Test: robotests still pass
Bug: 64092292
Change-Id: If9913d9ae08ee3e205ff324aaeeadc755ff1d23d