Set the bidi flags on Paints in DonutView
Previously, DonutView used Canvas.drawText with a Paint with default
bidi directions, which is LTR. This meant that even in RTL locales,
text was displayed assuming the direction of the paragraph the text
was appearing in was LTR. This caused an incorrect display of Arabic
percentages.

Now we set the Paint bidiFlags according to the Locale's direction.

Change-Id: Ic10228b8a23dc49de60246c37adfbaf7f8fd4e9e
Fixes: 63767043
Test: Manual (tested in ar-EG, ar-MA, fa-IR, ur-PK, and he-IL locales)