Fix BluetoothPairingDialogTest to not expect device name
The functional fix for b/62672248 in ag/2856166 removes the device name
from the text near the checkbox in the bluetooth pairing dialog, so we
need to stop checking for it in this robotest. However, that functional
fix needs to be in a separate CL from this fix for the robotest, because
the robotest doesn't exist on some of the older branches where the
functional fix needs to be backported.

Bug: 62672248
Test: RunSettingsRoboTests
Change-Id: I17ff3e4f13c20d58cc4027557f275d296ecccc95