CredentialStorage: Fix manual key installation.
Since the introduction of user-selectability of keys into KeyChain,
keys installed in KeyChain, by default, cannot be granted access
by the user. This is a secure default, and means keys manually
installed by the user should be explicitly marked as user-selectable.

This change marks the user-added alias as user-selectable, so it can be
selected, later on, by the user, from the Certificate Selection prompt.

Bug: 65624467
Test: Manually with CtsVerifier (KeyChainTest).
Change-Id: I618a06270a127a38b2aa3098fddfa518515bc1b1