Fix alignment in location requests screen
In Settings->Security & location->Location, the list of apps that have
recently requested location was moved up on the screen, and this made it
clear that the indentation of the app name was misaligned with the title
of Preference items that came after this. It turns out our custom layout
for AppPreference reserved a bit more space for icons than regular
Preference items.

While examining the layout file for location settings I also fixed a
small nit in a comment.

Fixes: 71056360

Test: Visual (go to Settings->Security & location->Location, and the app
names in the list should be indented at the same level as the Preference
items below)

Change-Id: I87e144c83cc2d13e322f65291415af037c405257