Add 50% and 100% to auto battery saver settings
The settings UI will be updated for P eventually, but let's add more
options to the settings for dogfooders until the UI update happens.

Also add the current value to the choices, so that if we push a number
that's not in the choices, the settings app will work properly.

Bug: 68769804
Bug: 71762373
Test: Manual test
Change-Id: I22fd14d43f4575885b4b8e068d3df1e65f60c9e1