Move bugreport storage enabling into OS.
It's convoluted to enable/disable bugreport storage based on the
"power menu" shortcut, and it's frustrating for dogfooders to have
to manually cycle this setting once to enable the provider.

So move the enable/disable logic directly into the OS, toggle it
when the broader DEVELOPMENT_SETTINGS_ENABLED changes, and update
to match state at device boot.

Test: builds, boots
Bug: 32540478
Change-Id: I8721dfbfc1dc2e19adb9cb5588f6b35f2872da9d