Update task affinity and parent activity for all exported activities.
This is needed to fix nav-up button for these activities.

- Change taskAffinity for homepage to "com.android.settings.root".
- Change taskAffinity for many activities to "com.android.settings", and
  set partent to homepage.
  -- General rule: if activity has not been exported, don't set
     affinity or parent.
  -- If activity is exported and can be launched with an action, set
     affinity and parent to homepage
  -- If activity is exported but expects additional parameters, don't
     set affinity or parent.
  -- If the activity affinitiy is com.android.setupwizard, don't touch
     it.

With this change, when a setting page is launched via intent (eg, from
QS), nav up button will be able to find parentActivity, and launch it in
a different task affinity. If the acitivity is a launched within
settings, nav-up has no change, it just navigates back to where it came from.

Change-Id: I3a2b64ed13d437f6818f6c7374e994b4192b6fa0
Fixes: 72337421
Fixes: 72829142
Test: manual
Test: atest