Handle Looper in PreferenceController
When indexing content in search, the preference controllers
are init'd on the background thread, thus you can't create a
handler without calling Looper.prepare().

We have PreferenceConttrollerContractTest to catch these
errors, but it was not running in presubmit.
Bug tracking: b/73242024

This caused the indexing thread to crash, thus
not adding any non-indexable keys for settings results.

Test: atest PreferenceControllerContractTest
Fixes: 73168342

Change-Id: I89fe596435d317bb3e3baebdaace97a67700d77e