Import translations. DO NOT MERGE
Merge translations of Bluetooth confirmation text

This is a manual merge of the translated strings from the master branch
of the 'bluetooth_pairing_shares_phonebook' string. The fix for this bug
requires these translations because the fix was to slightly reword the
confirmation text in the pairing dialog and stop substituting in the
device name.

Bug: 62672248
Test: manual
Change-Id: I43db31f1495e6873b6905896e6fa75fd88b4cb8d